requirements.txt
google-genai
pinecone
python-dotenv
pypdf
pymongo
pytz